refactor(blockifier): move BlockExecutionArtifacts to the blockifier crate
