+++ crates/blockifier/src/lib.rs
@@ -22,3 +22,5 @@ pub mod test_utils;
 pub mod transaction;
 pub mod utils;
 pub mod versioned_constants;
+// TODO(yael): create a new crate with the type in this module (and the types it depends on)
+pub mod execution_artifacts;
