+++ crates/starknet_batcher/Cargo.toml
@@ -30,6 +30,7 @@ validator.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
+blockifier = { workspace = true, features = ["testing"] }
 chrono = { workspace = true }
 futures.workspace = true
 mempool_test_utils.workspace = true
