+++ crates/starknet_batcher/src/block_builder.rs
@@ -10,6 +10,7 @@ use blockifier::blockifier::transaction_executor::{
 };
 use blockifier::bouncer::{BouncerConfig, BouncerWeights};
 use blockifier::context::{BlockContext, ChainInfo};
+use blockifier::execution_artifacts::BlockExecutionArtifacts;
 use blockifier::state::cached_state::CommitmentStateDiff;
 use blockifier::state::contract_class_manager::ContractClassManager;
 use blockifier::state::errors::StateError;
@@ -64,50 +65,6 @@ pub enum FailOnErrorCause {
     DeadlineReached,
     #[error("Transaction failed: {0}")]
     TransactionFailed(BlockifierTransactionExecutorError),
-}
-
-#[cfg_attr(test, derive(Clone))]
-#[derive(Debug, PartialEq)]
-pub struct BlockExecutionArtifacts {
-    pub execution_infos: IndexMap<TransactionHash, TransactionExecutionInfo>,
-    pub commitment_state_diff: CommitmentStateDiff,
-    pub visited_segments_mapping: VisitedSegmentsMapping,
-    pub bouncer_weights: BouncerWeights,
-    pub l2_gas_used: GasAmount,
-}
-
-impl BlockExecutionArtifacts {
-    pub fn address_to_nonce(&self) -> HashMap<ContractAddress, Nonce> {
-        HashMap::from_iter(
-            self.commitment_state_diff
-                .address_to_nonce
-                .iter()
-                .map(|(address, nonce)| (*address, *nonce)),
-        )
-    }
-
-    pub fn tx_hashes(&self) -> HashSet<TransactionHash> {
-        HashSet::from_iter(self.execution_infos.keys().copied())
-    }
-
-    pub fn state_diff(&self) -> ThinStateDiff {
-        // TODO(Ayelet): Remove the clones.
-        let storage_diffs = self.commitment_state_diff.storage_updates.clone();
-        let nonces = self.commitment_state_diff.address_to_nonce.clone();
-        ThinStateDiff {
-            deployed_contracts: IndexMap::new(),
-            storage_diffs,
-            declared_classes: IndexMap::new(),
-            nonces,
-            // TODO: Remove this when the structure of storage diffs changes.
-            deprecated_declared_classes: Vec::new(),
-            replaced_classes: IndexMap::new(),
-        }
-    }
-
-    pub fn commitment(&self) -> ProposalCommitment {
-        ProposalCommitment { state_diff_commitment: calculate_state_diff_hash(&self.state_diff()) }
-    }
 }
 
 /// The BlockBuilderTrait is responsible for building a new block from transactions provided by the
