+++ crates/starknet_batcher/src/test_utils.rs
@@ -3,6 +3,7 @@ use std::ops::Range;
 use async_trait::async_trait;
 use blockifier::blockifier::transaction_executor::VisitedSegmentsMapping;
 use blockifier::bouncer::BouncerWeights;
+use blockifier::execution_artifacts::BlockExecutionArtifacts;
 use blockifier::state::cached_state::CommitmentStateDiff;
 use indexmap::IndexMap;
 use starknet_api::executable_transaction::Transaction;
@@ -11,7 +12,7 @@ use starknet_api::test_utils::invoke::{executable_invoke_tx, InvokeTxArgs};
 use starknet_api::{class_hash, contract_address, nonce, tx_hash};
 use tokio::sync::mpsc::UnboundedSender;
 
-use crate::block_builder::{BlockBuilderResult, BlockBuilderTrait, BlockExecutionArtifacts};
+use crate::block_builder::{BlockBuilderResult, BlockBuilderTrait};
 use crate::transaction_provider::{NextTxs, TransactionProvider};
 
 // A fake block builder for validate flow, that fetches transactions from the transaction provider
@@ -70,23 +71,21 @@ pub fn test_txs(tx_hash_range: Range<usize>) -> Vec<Transaction> {
         .collect()
 }
 
-impl BlockExecutionArtifacts {
-    pub fn create_for_testing() -> Self {
-        // Use a non-empty commitment_state_diff to make the tests more realistic.
-        Self {
-            execution_infos: IndexMap::default(),
-            commitment_state_diff: CommitmentStateDiff {
-                address_to_class_hash: IndexMap::from_iter([(
-                    contract_address!("0x7"),
-                    class_hash!("0x11111111"),
-                )]),
-                storage_updates: IndexMap::new(),
-                class_hash_to_compiled_class_hash: IndexMap::new(),
-                address_to_nonce: IndexMap::from_iter([(contract_address!("0x7"), nonce!(1_u64))]),
-            },
-            visited_segments_mapping: VisitedSegmentsMapping::default(),
-            bouncer_weights: BouncerWeights::empty(),
-            l2_gas_used: GasAmount::default(),
-        }
+pub fn testing_block_execution_artifacts() -> BlockExecutionArtifacts {
+    // Use a non-empty commitment_state_diff to make the tests more realistic.
+    BlockExecutionArtifacts {
+        execution_infos: IndexMap::default(),
+        commitment_state_diff: CommitmentStateDiff {
+            address_to_class_hash: IndexMap::from_iter([(
+                contract_address!("0x7"),
+                class_hash!("0x11111111"),
+            )]),
+            storage_updates: IndexMap::new(),
+            class_hash_to_compiled_class_hash: IndexMap::new(),
+            address_to_nonce: IndexMap::from_iter([(contract_address!("0x7"), nonce!(1_u64))]),
+        },
+        visited_segments_mapping: VisitedSegmentsMapping::default(),
+        bouncer_weights: BouncerWeights::empty(),
+        l2_gas_used: GasAmount::default(),
     }
 }
